#25 refactor: Move database and repository related code into a new Data project

+++ OrderService.Entities/Product.cs
@@ -1,3 +1,3 @@
 namespace OrderService.Entities;
 
-public record Product(Guid Id, string Name, double Price);
+public record Product(Guid Id, string Name, decimal Price);
